# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BLC API Docs",
        "description": "BLC Swagger OpenAPI documentation",
        "contact": {
            "name": "BLC Swagger API",
            "email": "mukelele@mukelele.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/privacy-policy": {
            "get": {
                "tags": [
                    "Home"
                ],
                "summary": "Get privacy policy page",
                "operationId": "aafcf2253582bb65b1af61d52333b783",
                "responses": {
                    "200": {
                        "description": "Success response"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/terms-and-conditions": {
            "get": {
                "tags": [
                    "Home"
                ],
                "summary": "Get terms and conditions page",
                "operationId": "288f8add9e490f635451bd6e95115df0",
                "responses": {
                    "200": {
                        "description": "Success response"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}